rename computeCompositionPropertyValues to getCompositionRenderValues
